requirements.txt
pytest==8.0.0
httpx==0.26.0
dependency-injector==4.41.0
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.12.0